Assignment 3 solution
Quality of UI (15 points): How impressive your UI is (Look and Feel). 
 
Task 0: Front-page Activity (15 points). Use the Cover Page from the
Assignment 2. In the last few assignments, we have been using the menu
on the Action Bar for our navigation. In this assignment, we will use
buttons as well as using the menu on the Action Bar or App Bar for
navigating other task activities. You need to build an Activity (called
Front-page), which loads a front-page Fragment. Inside this Fragment,
you need to have the following buttons (you can decide the layout and
feel free to add other decoration items):
•	 “About me” button: when it is clicked, your “about me” Fragment will
be loaded. You need to change the orientation of your device to check
whether it works properly.
•	 A button for Task 2 and Task 3: when it is clicked, the
corresponding Activity will come up (same operation with the menu on
the Action Bar or App Bar).
 
Task 1: Build a movie fragment (15 points). You need to implement a
Fragment to display the information for a movie. Our movie data contain
many fields, and you should display at least the following information:
•	Title
•	Description
•	Image
•	Year
•	Stars
•	Rating (Don’t use TextView here; find the most suitable UI widget to
display this information)
•	You are responsible for designing the layout of this task. Make it
looks nice. This layout will be reused again.
•	There is no need to demonstrate this task, because the fragment will
be used in the next two tasks, and we will grade your fragment based on
what we see in the next two tasks.
 
Task 2: ViewPager Activity (30 points). Implement
a ViewPager activity. Each page of the ViewPager is a fragment that
displays the details of a movie. Users can swipe left or right to go
the previous or next movie. You should satisfy the following
requirements:
•	You should use the fragment developed in Task 1.
•	You need to have a tab (one tab for each movie). When user clicks on
a tab, the corresponding movie should be displayed.
 
Task 3: Master/Detail Flow (25 points).  This task consists of two
fragments: Master and Detail. Implement an Activity that initially load
a Master fragment with 1 TextView and 2 Buttons (could be image
buttons) (this fragment is called Movie_Choice fragment); each of the
buttons corresponds to the index of a movie. When the increase button
is clicked, increase the index by the maximum (# of movies) and when
the decrease button is clicked, decrease the index by 0 (cannot be
negative). Then your app should behave like the following:
•	On Handset with a small screen, the Detail fragment showing the
details of the movie will be loaded into the same activity, replacing
the Movie_Choice fragment. When the “back” button of the phone is
clicked, the Movie_Choice fragment with same index value will come back.
•	On Tablet with a large screen, the fragment showing the details of
the movie will be loaded into the same activity, but both the
Movie_Choice fragment and the detail fragment will be shown side by
side.