Assignment 4 solution
Look and Feel (15 points): The quality of your UI will be evaluated in
this homework. 
 
Task 0: Front-page Activity (5 points). You can reuse the one built
from the last homework. You need to build an Activity (called front
page), which loads a front-page Fragment. Inside this Fragment, you
need to have the following buttons (you can decide the layout and feel
free to add other decoration items):
•	 “About me” button: when it is clicked, your “about me” Fragment will
be loaded. You need to change the orientation of your device to check
whether it works properly.
•	 A button for Task 1: when it is clicked, the
corresponding Activity will come up (same operation with the menu on
the Action Bar or App Bar).
 
Task 1 (80): Please load the movie data that we provide to you into
a RecyclerView. Each row of the RecyclerView should contain at least
the following information: the image of movie, title, description,
rating, and a CheckBox. You can design your own layout. There should
be four buttons above the RecyclerView. You need to implement the
following functionalities:
1.	(20) Get the RecyclerView to work; each row should be a CardView.
2.	(5) When the CheckBox is clicked, its status will be toggled.
However, when the row is clicked, the CheckBox will not be toggled.
3.	(5) When the “Select All” button is clicked, all the CheckBoxes will
be checked. When the “Clear All” button is clicked, all
the CheckBoxes will be unchecked.
4.	(5) When the “Delete” button is clicked, all the selected movies
will be deleted from the list. Users should be able to see the
animation effect.
5.	(10) When a list item is long clicked, that item will be duplicated
and placed right below the current item (i.e., a new movie item with
duplicated content will appear). Users should be able to see the
animation effect.
6.	(10) When a list item is clicked, a Fragment will be loaded into the
current activity, displaying the details of the selected movie.
7.	(5) When the phone’s orientation changes, the selected movies should
stay selected.
8.	(10) When the "Sort" button is clicked, all movies will be sorted by
the year (the most recent movie should be the top of the list). The top
5 movies and the bottom 5 movies should have different layouts from
other middle movies (There should be 3 layouts for each group). 
9.	(5) Try an Item animation
from wasabeef library https://github.com/wasabeef/recyclerview-animators
10.	(5) Try an adapter animation from wasabeef library