Assignment 5 solution
Look and Feel (15 points): The quality of your UI will be evaluated.
Task 1 Navigation Drawer (20 points). In this lab, we will use the
navigation drawer as the top-level navigation. You need to have the
followings on your drawer:
1.	 A nice-looking banner (header)
2.	At least 3 items, each containing an icon and a title. One item will
bring out the "about me" fragment, and the other two items will bring
out the Activities for Task 2 and Task 3, respectively.
 
Task 2 (50 points): Please implement an Activity, which loads
a RecyclerView Fragment (Use your RecyclerView Fragment and Adapter
from Assignment 4):
1.	(10 points) You have to use Toolbar for your Activity's action bar.
On the Toolbar, you need to use an Imageview to display a picture, and
an TextView to display Title (For example, title can be "Movie List"
in RecyclerView fragment, and in Detail Fragment, title can be the name
of the selected movie). 
2.	(10 points) The RecyclerView Fragment’s action bar contains
a SearchView. After the user types a string in the text
field, RecyclerView should scroll to the first movie item that contains
the string (non case-sensitive).
3.	inside Detail View Fragment (10 points) When a user clicks a list
item, the detail view (a fragment) of the selected movie will be loaded
into the same Activity. There is a Share action provider on the action
bar of the fragment. This action provider should allow users to send
some information about the selected movie to other apps (e.g.
message app). When you tap a certain text on the
description TextView, popup CAB for copy and paste.

4.	 (10 points) When a user long clicks an item, the contextual action
bar should appear, offering two choices (duplicate the item or delete
the item). You need to implement the duplication and deletion.
5.	 (10 points) Each item should have an ImageView (see the 3 vertical
dots in example); when the ImageView is clicked, a popup menu will
appear, offering two choices (duplicate the item or delete the item).
You need to implement the duplication and deletion.
 
Task 3 (15 points): Please implement another Activity, which does
exactly the same thing as in Task 2 (i.e., you can reuse the
same RecyclerView Fragment), but you need to implement the followings
1.	You have to add a standalone Toolbar to your app, and place it at
the bottom of the Activity. Therefore, your app has two Toolbars, one
used as the Action bar (i.e. app bar), and the other is this
standalone Toolbar. 
2.	On these Toolbars, place items to do sorting movie data by name and
year. For example, when you trigger sort-by-name item on the action bar
(top), your movie list will be sorted and the sorted list will be
displayed. And sort-by-year item is on the standalone toolbar (bottom).

3.	There should be buttons somewhere to hide and unhide this
standalone Toolbar.