assignment 2 solution
•	Theme (10 points). You should use different themes for each Task.
There should be 4 different themes and they should be quite different
visually (Color, Font, Background, etc.)  
 
•	Taks 1: Cover page (20 points). On the MainActivity for the front
page, you should display the following information: your name, the
course that you have enrolled in (e.g. CIS400 or CIS600), and a photo
of yourself. This activity should also have a menu
(using Toolbar, ActionBar is deprived), one menu item for each task.
When a menu item is clicked, the Activity that implements the
corresponding task should come up. Namely, each of the tasks should
have its own activity and layout. This cover page will be used as the
front page of the following assignments, so it is worth to invest time
to make this page look good. You can choose any layout for this task.

•	Task 2: Layout Only (10 points): Implement the following interface
using LinearLayout.

•	Task 3: Simple Calculator - Design and UI Event Handling (30
points): Implement a simple Integer calculator with your choice of
Layout (Design -- set of Linear Layouts, Grid Layout, or Table Layout )
with 10 number buttons(0 to 9), 6 operator(+,-,/,*,%, =) button, and
1 TextView to display user input and calculated output when the (=)
button is clicked. 

•	Task 4: Movie List - Control and Events (30 points). In this
activity, you need to have an information table, a SeekBar and
an ImageView. The information table display all information except
image of the selected movie. The ImageView should display the first
movie image from the Movie data provided to you. The mark of
the SeekBar should be placed in the middle initially. You need to do
the followings:
o	When the user slides the SeekBar, the size of the image should change
accordingly.
o	When the user long-presses the image, the mark of the SeekBar goes to
the middle (i.e., the initial position), and the size of the image
should change accordingly.
o	When the user clicks (short press) the image, a Toast message and a
snack bar message should both be displayed.
o	When the user swipes to the right, the previous movie picture  and
its information should be loaded (if the current picture is the first
one, then it stays the same). 
o	When the user swipes to the left, the next movie picture and its
information should be loaded (if the current picture is the last one,
then it stays the same).