Assignment 1 solution
Write a Hello-World app that satisfies the following requirements

(20 points) Change the app icon to a picture of your choice.
(20 points) Add a TextView to your Activity. This TextView should
display your full name.
(20 points) Add another TextView. When the phone is in the portrait
mode, it shows “Hello Syracuse”. When the phone is in the landscape
mode, it shows “Hello Orangemen”.
(20 points) Add an ImageView to your Activity. It displays an image of
your choices. When the phone is in the landscape mode and portrait
mode, the images should be different.
Be careful of your image sizes! If your image size is too big, you
would have an error. Then reduce the size and resolution of your image.
(20 points) Add a Button to your interface. Your app does not need to
do anything when the button is clicked.
Use string definition defined in strings.xml for the text of your
Button. On the Button, "Do not Click" is shown.