assignment 8 solution
You should modify the RecyclerView fragment that you wrote in the past.
Since we are not asking you to develop new UI, we will not give "look
and feel" score for this homework. You need to do the following tasks:

Due to the complexity of this lab, the video length is increased to 4
minutes.

You can use the MyUtility.java that Dr. Du wrote. This class provides
methods to get data from and update data to servers via HTTP GET and
POST requests.

Task 1 (40 points). Load your movie data from a URI such as
www.example.com/movies/ (feel free to use your own URI). Each movie
entry has a field called “url”, which is where you should download the
picture for the movies. You can’t use any local image; all the movie
images must be downloaded from the Internet. You need to provide
evidences during your demo that you are indeed getting everything from
outside. For example, you can disconnect your WiFi, kill your HTTP
server, etc. Your app should show the differences.

Task 2 (10 points). On the action bar of RecyclerView, please include a
“search” action view, i.e. SearchView (we covered that in our Action
Bar lecture). When users type a number in the text field, the
RecyclerView will refresh and display all the movies with ratings
larger than or equal to the number typed in the SearchView field. The
URI that you use to get the movie list is
www.example.com/movies/rating/8.0.  The number 8.0 should be replaced
by whatever is typed, but the prefix www.example.com/movies/rating/
should be the same. You should write the server-side code, which
returns an array of json objects.

Task 3 (20 points). When a movie is selected, the detail fragment
should be loaded, which displays the details of the movie, including
its name, description, year, length, stars, directors, rating, and
images (you can reuse the fragment that you built in the past
homework). The details of the movie are not included in the json array
received in Task 1; instead, you need to fetch the details from your
server. Your URI for the movie details should look like the following:
www.example.com/movies/id/titanic (there is no “/” at the end of the
URI). Your server-side code should return a single json object.

Task 4 (30). In the RecyclerView homework, you have implemented a popup
menu for "duplicate" and "delete" movies. In this task, when the
"duplicate" menu item is clicked, a duplicated copy of the selected
movie will be inserted into the server-side database (please change the
id of the movie, e.g., by attaching "_new" after the existing id). When
the "delete" menu item is clicked, the selected movie is actually
deleted. You need to provide the following evidences to show that the
databases are indeed modified.
Go to XAMPP and show the database from there. Point out the added
movies and the evidence for the deleted movies from the database.
Kill your app, and run it again. Point out the added movies and the
evidence for the deleted movies in the RecyclerView.

Submission Requirements: You should submit the followings:
All your PHP programs
Your app’s apk file
Video demonstrating requirements